requirements.txt
streamlit>=1.34
chromadb>=0.5.0
sentence-transformers>=2.6.1
transformers>=4.42.0
torch>=2.1.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
tqdm>=4.66.0
pydantic>=2.6.0
gdown
faiss-cpu
